+++ example/src/main/scala/geotrellis/server/example/ndvi/NdviService.scala
@@ -1,36 +1,22 @@
 package geotrellis.server.example.ndvi
 
 import geotrellis.server._
-import TmsReification.ops._
-import com.azavea.maml.util.Vars
+import geotrellis.raster._
+import geotrellis.raster.render._
 import com.azavea.maml.ast._
 import com.azavea.maml.ast.codec.tree._
 import com.azavea.maml.eval._
+
 import org.http4s._
 import org.http4s.dsl.Http4sDsl
 import org.http4s.circe._
-import io.circe._
-import io.circe.parser._
-import io.circe.syntax._
-import cats._
+import _root_.io.circe._
+import _root_.io.circe.parser._
+import _root_.io.circe.syntax._
 import cats.data._
 import Validated._
-import cats.implicits._
 import cats.effect._
 import com.typesafe.scalalogging.LazyLogging
-import geotrellis.raster._
-import geotrellis.raster.render._
-
-import scala.math._
-import java.net.URI
-import java.util.{NoSuchElementException, UUID}
-
-import geotrellis.server.cog.CogNode
-import geotrellis.server.gdal.GDALNode
-
-import scala.util.Try
-import scala.collection.mutable
-
 
 class NdviService[Param](
   interpreter: BufferingInterpreter = BufferingInterpreter.DEFAULT
@@ -70,11 +56,8 @@ class NdviService[Param](
   // http://0.0.0.0:9000/{z}/{x}/{y}.png
   def routes: HttpRoutes[IO] = HttpRoutes.of {
     // Matching json in the query parameter is a bad idea.
-    case req @ GET -> Root / IntVar(z) / IntVar(x) / IntVar(y) ~ "png" /*:? RedQueryParamMatcher(red) +& NirQueryParamMatcher(nir)*/ =>
-      val paramMap = Map(
-        "red" -> CogNode(new java.net.URI("file:///Users/daunnc/subversions/git/github/geotrellis-landsat-tutorial/data/r-g-nir.tif"), 0, None).asInstanceOf[Param],
-        "nir" -> CogNode(new java.net.URI("file:///Users/daunnc/subversions/git/github/geotrellis-landsat-tutorial/data/r-g-nir.tif"), 2, None).asInstanceOf[Param]
-      )
+    case req @ GET -> Root / IntVar(z) / IntVar(x) / IntVar(y) ~ "png" :? RedQueryParamMatcher(red) +& NirQueryParamMatcher(nir) =>
+      val paramMap = Map("red" -> red, "nir" -> nir)
 
       eval(paramMap, z, x, y).attempt flatMap {
         case Right(Valid(mbtile)) =>
