+++ example/src/main/scala/geotrellis/server/example/ndvi/gdal/NdviService.scala
@@ -1,27 +1,23 @@
 package geotrellis.server.example.ndvi.gdal
 
 import geotrellis.server._
-import TmsReification.ops._
-import com.azavea.maml.util.Vars
-import com.azavea.maml.ast._
-import com.azavea.maml.ast.codec.tree._
-import com.azavea.maml.eval._
-import org.http4s._
-import org.http4s.dsl.Http4sDsl
-import org.http4s.circe._
-import io.circe._
-import io.circe.parser._
-import io.circe.syntax._
-import cats._
-import cats.data._
-import Validated._
-import cats.implicits._
-import cats.effect._
-import com.typesafe.scalalogging.LazyLogging
 import geotrellis.raster._
 import geotrellis.raster.render._
 import geotrellis.server.gdal.GDALNode
+import com.azavea.maml.ast._
+import com.azavea.maml.ast.codec.tree._
+import com.azavea.maml.eval._
 
+import org.http4s._
+import org.http4s.dsl.Http4sDsl
+import org.http4s.circe._
+import _root_.io.circe._
+import _root_.io.circe.parser._
+import _root_.io.circe.syntax._
+import cats.data._
+import Validated._
+import cats.effect._
+import com.typesafe.scalalogging.LazyLogging
 
 class NdviService[Param](
   interpreter: BufferingInterpreter = BufferingInterpreter.DEFAULT
@@ -61,14 +57,8 @@ class NdviService[Param](
   // http://0.0.0.0:9000/{z}/{x}/{y}.png
   def routes: HttpRoutes[IO] = HttpRoutes.of {
     // Matching json in the query parameter is a bad idea.
-    case req @ GET -> Root / IntVar(z) / IntVar(x) / IntVar(y) ~ "png" /*:? RedQueryParamMatcher(red) +& NirQueryParamMatcher(nir)*/ =>
-
-      println(s"geotrellis.contrib.vlm.gdal.GDAL.cache.size: ${geotrellis.contrib.vlm.gdal.GDAL.cache.asMap.size}")
-
-      val paramMap = Map(
-        "red" -> GDALNode(new java.net.URI("/Users/daunnc/subversions/git/github/geotrellis-landsat-tutorial/data/r-g-nir.tif"), 0, None).asInstanceOf[Param],
-        "nir" -> GDALNode(new java.net.URI("/Users/daunnc/subversions/git/github/geotrellis-landsat-tutorial/data/r-g-nir.tif"), 2, None).asInstanceOf[Param]
-      )
+    case req @ GET -> Root / IntVar(z) / IntVar(x) / IntVar(y) ~ "png" :? RedQueryParamMatcher(red) +& NirQueryParamMatcher(nir) =>
+      val paramMap = Map("red" -> red, "nir" -> nir)
 
       eval(paramMap, z, x, y).attempt flatMap {
         case Right(Valid(mbtile)) =>
